Fix mishap in flutter integration tests.. if ariTester does not have exec privilege, equity plan can't be updated, and onboarding can't run without an equity plan. Now, ariTester roles are set before every test - flutter or server.
