+++ webServer/tests/testSaveDynamo.js
@@ -56,7 +56,6 @@ async function runTests( flutterTest ) {
     let success = false;
 
     // turn this on when not testing in parts
-
     success = execAWS_CLI( "CEPEQs", flutterTest );
     testStatus = tu.checkEq( success, true, testStatus, "save PEQ Table" );
